Plots data processing with VTK elimination. Workspace datasets edition. Map widget tools.
Workspace datasets edition
- Implemented methods for displaying file description (not concluded).
- Moved CDatasetBrowserControl class to a new file.
- CHANGED:  method "DatasetChanged", "FillVarList".
- ADDED:       method "ClearVarList".
- Development of methods for displaying "File description" and "Variables list".
- Added icon to color button.
- Removed axis options (change color, opacity, pattern, width).

Plots data
- Processed all plot types data. Eliminated all VTK dependencies.
- Included methods on QwtArrayPlotData and CCurve to calculate ranges accurately. Also included very valuable comments that describe the same obscure versions on the old brat display code.
- Correction in velocity maps data processing: solved v3 bug with coda initialization.
- Corrections in XY and LatLon plots data.
- Corrections in 2D multi-frame plots.

Map widget tools
- map measure and grid tools
- map selections tool

Misc
- Corrected bug assertion in release versions
- Branch adjustments to all platforms
- Updated all supported IDEs projects

+++ display/vtkInteractorStyleWPlot.h
@@ -98,7 +98,7 @@ public:
   void RestoreState(vtkInteractorStyleWPlot*, int32_t index);
   void ClearStates();
   bool HasState() {return CountState() > 0;};
-  uint32_t CountState() {return this->States.size();};
+  uint32_t CountState() {return (uint32_t)this->States.size();};
 
   vtkWPlotState* GetState(int32_t index);
